\docType{methods}
\name{as.character,Pattern-method}
\alias{as.character,Pattern-method}
\title{as.character.Pattern}
\usage{
\S4method{as.character}{Pattern}(x, ...)
}
\arguments{
  \item{x}{object to be coerced or tested.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\description{
as.character.Pattern
}
\keyword{internal}

