% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docopt-package.R
\docType{package}
\name{docopt-package}
\alias{docopt-package}
\title{Docopt command line specification}
\description{
docopt helps you to define an interface for your command-line app, and
automatically generate a parser for it.
}
\details{
For more information see http://docopt.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/docopt/docopt.R}
  \item Report bugs at \url{https://github.com/docopt/docopt.R/issues}
}

}
