% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Pattern.R
\docType{methods}
\name{as.character,Pattern-method}
\alias{as.character,Pattern-method}
\title{as.character.Pattern}
\usage{
\S4method{as.character}{Pattern}(x, ...)
}
\description{
as.character.Pattern
}
\keyword{internal}

