\docType{package}
\name{docopt-package}
\alias{docopt-package}
\title{docopt command line specification}
\description{
  docopt helps you to define an interface for your
  command-line app, and automatically generate a parser for
  it.
}
\details{
  For more information see \link{http://docopt.org}
}

