% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{docopt-package}
\alias{docopt-package}
\title{Docopt command line specification}
\description{
docopt helps you to define an interface for your command-line app, and
automatically generate a parser for it.
}
\details{
For more information see http://docopt.org
}

